requirements.txt
tortoise-orm
aiohttp
fastapi
uvicorn[standart]
pydantic
pyjwt